Remove WherebyRoomActivity flow and event broadcast receiver
