+++ java/app/build.gradle
@@ -37,5 +37,5 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 
     // Whereby SDK
-    implementation("com.github.whereby:android-sdk:0.0.6@aar") { transitive = true }
+    implementation("com.github.whereby:android-sdk:0.0.8@aar") { transitive = true }
 }
